requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
pypdf>=3.17.0
chromadb>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
ctransformers>=0.2.0
sentence-transformers>=2.2.0

# Speech functionality dependencies
openai-whisper>=20231117
gTTS>=2.5.0
streamlit-webrtc>=0.47.1
av>=10.0.0
librosa>=0.10.0
transformers>=4.36.0

# Additional dependencies
requests>=2.31.0
python-multipart>=0.0.6